fix baithi

+++ app/Http/Controllers/admin/LopHocController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\CTHoaDon;
 use App\Models\KhoaHoc;
 use App\Models\LopHoc;
 use Illuminate\Http\Request;
@@ -31,7 +32,11 @@ class LopHocController extends Controller
                 ->where('MAGV', session('login')->ID)->get();
         else
             $khoahoc = KhoaHoc::where('TRUCTUYEN', '=', 1)->get();
-        return view('admin.lophoc.create', ['khoahoc' => $khoahoc]);
+        $check = 1;
+        if (count($khoahoc) > 0)
+            return view('admin.lophoc.create', compact('khoahoc', 'check'));
+        $check = 0;
+        return view('admin.lophoc.create', compact('khoahoc', 'check'));
     }
 
     public function store(Request $request)
@@ -68,9 +73,12 @@ class LopHocController extends Controller
     public function delete($id)
     {
         $lophoc = LopHoc::find($id);
-        $lophoc->delete();
-
-        return redirect('admin/lophoc/')->with('thongbao', 'Xóa thành công!');
+        $hoadon = CTHoaDon::where('MALH', $id)->get();
+        if (count($hoadon) == 0) {
+            $lophoc->delete();
+            return redirect('admin/lophoc/')->with('thongbao', 'Xóa thành công!');
+        } else
+            return redirect('admin/lophoc/')->with('thongbao', 'Lớp học đã có người mua. Không thể xóa lớp học này!');
     }
 
     public function search(Request $request)
